2023-07-03 update v1.3.0

+++ app/Http/Controllers/Api/CommentController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Models\Blog;
 use App\Models\Comment;
 use Illuminate\Http\Request;
 use Illuminate\Support\Carbon;
@@ -11,8 +12,12 @@ use Illuminate\Support\Facades\Auth;
 class CommentController extends Controller
 {
     public function index(Request $request){
-        $comment=Comment::where('BlogID',$request->id)->get();
-        return response()->json(['result'=>true,'comment' => $comment]);
+        $comment=Comment::where('BlogID',$request->id)->with('blog','user')->get();
+        foreach($comment as $item){
+            $item['permission']=Auth::user() && Auth::user()->id==$item->UserID ?true:false;
+        }
+        $permission=Auth::user() && Auth::user()->id==Blog::find($request->id)->UserID ?true:false;
+        return response()->json(['result'=>true,'comment' => $comment,'permission'=>$permission]);
     }
 
     public function store(Request $request){
@@ -21,6 +26,9 @@ class CommentController extends Controller
         $comment->BlogID=$request->BlogID;
         if($request->CommentID!=='null'){
             $comment->CommentID=$request->CommentID;
+        }
+        if(Auth::user()){
+            $comment->UserID=Auth::user()->id;
         }
         $comment->date=Carbon::now();
         $query=$comment->save();
@@ -44,8 +52,9 @@ class CommentController extends Controller
 
     public function delete(Request $request){
         $user=Auth::user();
-        $comment=Comment::find($request->CommentID);
-        if($comment && $user){
+        $comment=Comment::with('blog')->find($request->CommentID);
+        $permission=Auth::user() && Auth::user()->id==$comment->blog->UserID ?true:false;
+        if($comment && $user && $permission){
             $query=$comment->delete();
             if($query){
                 return response()->json(['result'=>true]);
